+++ .kokoro/samples-test.sh
@@ -20,6 +20,10 @@ set -eo pipefail
 mkdir -p $NPM_CONFIG_PREFIX
 npm config -g ls || npm i -g npm@`npm --version`
 
+# Ensure the npm global directory is writable, otherwise rebuild `npm`
+mkdir -p $NPM_CONFIG_PREFIX
+npm config -g ls || npm i -g npm@`npm --version`
+
 # Setup service account credentials.
 export GOOGLE_APPLICATION_CREDENTIALS=${KOKORO_GFILE_DIR}/secret_manager/long-door-651-kokoro-system-test-service-account
 export GCLOUD_PROJECT=long-door-651