+++ system-test/storage.ts
@@ -435,9 +435,9 @@ describe('storage', function () {
           resumable: false,
         });
         const [metadata] = await file.getMetadata();
-        const encyrptionAlgorithm =
+        const encryptionAlgorithm =
           metadata.customerEncryption?.encryptionAlgorithm;
-        assert.strictEqual(encyrptionAlgorithm, 'AES256');
+        assert.strictEqual(encryptionAlgorithm, 'AES256');
       });
 
       it('should set custom encryption in a resumable upload', async () => {
@@ -447,9 +447,9 @@ describe('storage', function () {
           resumable: true,
         });
         const [metadata] = await file.getMetadata();
-        const encyrptionAlgorithm =
+        const encryptionAlgorithm =
           metadata.customerEncryption?.encryptionAlgorithm;
-        assert.strictEqual(encyrptionAlgorithm, 'AES256');
+        assert.strictEqual(encryptionAlgorithm, 'AES256');
       });
 
       it('should make a file public during the upload', async () => {
@@ -2778,7 +2778,7 @@ describe('storage', function () {
         });
       });
 
-      it('should download from the encrytped file', async () => {
+      it('should download from the encrypted file', async () => {
         const [contents] = await file.download();
         assert.strictEqual(contents.toString(), 'secret data');
       });